chore(api): refactor user module
