+++ .moon/tasks.yml
@@ -39,13 +39,25 @@ tasks:
   dev:
     command: "cargo watch"
     local: true
+    options:
+      cache: false
   build:
     command: "cargo build"
+    options:
+      cache: false
   release:
     command: "cargo build --release"
+    options:
+      cache: false
   test:
     command: "cargo nextest run"
+    options:
+      cache: false
   format:
     command: "cargo fmt"
+    options:
+      cache: false
   lint:
     command: "cargo clippy"
+    options:
+      cache: false
