+++ crates/server/src/router/routes/login/post.rs
@@ -1,5 +1,5 @@
 use crate::html::HtmlTemplate;
-use api::{user::LoginPayload, KICKBASE};
+use api::{user::login::LoginPayload, KICKBASE};
 use askama::Template;
 use askama_axum::IntoResponse;
 use axum::{http::HeaderValue, response::Response, Json};
